moved outdoor map to new page
