+++ SD-K9/package.json
@@ -26,6 +26,7 @@
     "@ionic-native/splash-screen": "^5.21.6",
     "@ionic-native/status-bar": "^5.21.6",
     "@ionic/angular": "^4.7.1",
+    "@svgdotjs/svg.js": "^3.0.16",
     "@types/hammerjs": "^2.0.36",
     "cordova-android": "^8.1.0",
     "cordova-browser": "^6.0.0",
@@ -35,7 +36,6 @@
     "ionic-pullup": "^4.0.0-beta.3",
     "karma-firefox-launcher": "^1.3.0",
     "rxjs": "~6.5.1",
-    "@svgdotjs/svg.js": "^3.0.16",
     "tslib": "^1.11.0",
     "zone.js": "~0.9.1"
   },
