+++ SD-K9/src/app/app-routing.module.ts
@@ -12,6 +12,10 @@ const routes: Routes = [
     path: 'appsettings',
     loadChildren: () => import('src/app/modals/appsettings/appsettings.module').then(m => m.AppsettingsPageModule)
   },
+  {
+    path: 'outdoor',
+    loadChildren: () => import('./pages/outdoor/outdoor.module').then( m => m.OutdoorPageModule)
+  },
 ];
 
 @NgModule({
